test: revamp sysfs tree dump test
In order to extend the sysfs tree dump tests, let's improve it.
